+++ src/routes/login.tsx
@@ -1,7 +1,6 @@
 import React, { useLayoutEffect, useState } from 'react'
 import Modal from '../components/common/modal'
 import LoginModal from '../components/auth/loginModal'
-import SignUpModal from '../components/auth/signUpModal'
 import { FaCircleArrowRight } from 'react-icons/fa6'
 import { LoginModalType } from '../types/app'
 import FindPwdModal from '../components/auth/findPwdModal'
